move vault details down the tree
